api/requirements.txt
fastapi
mangum
groq
python-dotenv